requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
opencv-python-headless
transformers
timm
python-multipart